+++ components/atoms/Text.js
@@ -1,8 +1,0 @@
-import React from 'react'
-import { Text as RebassText } from 'rebass/styled-components'
-
-const Text = props => (
-  <RebassText {...props} />
-)
-
-export default Text