+++ src/components/Text.tsx
@@ -1,0 +1,6 @@
+import React, { FC } from 'react'
+import { Text as RebassText, TextProps } from 'rebass/styled-components'
+
+const Text: FC<TextProps> = (props) => <RebassText {...props} />
+
+export default Text