+++ src/components/icons/CategoryCodeGRP.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import Icon from 'react-icon-base'
+const CategoryCodeGRP = (props: any) => (
+  <Icon viewBox="0 0 16 16" {...props}>
+    <g>
+  <title>CategoryCodeGRP</title>
+  <g>
+    <path d="M8.906,8.328h-4a2.534,2.534,0,0,0-2.527,2.527V12.57H8.382a3.234,3.234,0,0,1,.524-4.242Z"/>
+    <circle cx="7.225" cy="5.26" r="2.572"/>
+    <path d="M11.061,8.183a2.565,2.565,0,1,0,2.564,2.564A2.564,2.564,0,0,0,11.061,8.183Zm1.563,2.965h-1.1v1.1H10.6v-1.1H9.5V10.22h1.1v-1.1h.929v1.1h1.1Z"/>
+  </g>
+</g>
+  </Icon>
+)
+export default CategoryCodeGRP